+++ Visualizer/Engine/Style DSL/Sundex.Style.DSL/Abstract/Values/NumberValue.cs
@@ -1,6 +1,6 @@
 namespace Sundex.Style.DSL.Abstract.Values;
 
-public record NumberValue(double Value, string Unit) : IStyleValue
+public record NumberValue(float Value, string Unit) : IStyleValue
 {
     object IStyleValue.Value => Value;
     public override string ToString() => $"{Value}{Unit}";
